+++ src/function.rs
@@ -1,29 +1,58 @@
 use std::sync::Arc;
 
-use crate::{DataSources, Error, Transaction, DATA_SOURCES, DEFAULT_DATA_SOURCE_NAME};
+use crate::{DataSources, Error, Transaction, DATA_SOURCES, DEFAULT_DATA_SOURCE};
 
+#[inline(always)]
 pub async fn default_txn() -> Result<Transaction, Error> {
-    data_sources()?.current_txn(DEFAULT_DATA_SOURCE_NAME).await
+    current_txn(DEFAULT_DATA_SOURCE).await
 }
 
 pub async fn current_txn(name: &str) -> Result<Transaction, Error> {
-    data_sources()?.current_txn(name).await
+    let txn = data_sources()?
+        .get(name)
+        .ok_or_else(|| Error::NotFoundDataSource { name: name.into() })?
+        .current_txn()
+        .await?;
+
+    Ok(txn)
 }
 
-pub async fn new_txn(name: &str) -> Result<Transaction, Error> {
-    data_sources()?.new_txn(name).await
+pub async fn create_txn(name: &str) -> Result<Transaction, Error> {
+    let txn = data_sources()?
+        .get(name)
+        .ok_or_else(|| Error::NotFoundDataSource { name: name.into() })?
+        .create_txn()
+        .await?;
+
+    Ok(txn)
 }
 
-pub async fn commit(name: &str) -> Result<(), Error> {
-    data_sources()?.commit(name).await
+pub async fn commit(txn: Transaction) -> Result<(), Error> {
+    let data_sources = data_sources()?;
+
+    let source = data_sources
+        .get(&txn.name)
+        .ok_or_else(|| Error::NotFoundDataSource {
+            name: txn.name.as_ref().into(),
+        })?;
+
+    source.commit(txn).await
 }
 
-pub async fn rollback(name: &str) -> Result<(), Error> {
-    data_sources()?.rollback(name).await
+pub async fn rollback(txn: Transaction) -> Result<(), Error> {
+    let data_sources = data_sources()?;
+
+    let source = data_sources
+        .get(&txn.name)
+        .ok_or_else(|| Error::NotFoundDataSource {
+            name: txn.name.as_ref().into(),
+        })?;
+
+    source.rollback(txn).await
 }
 
 pub fn data_sources() -> Result<Arc<DataSources>, Error> {
     DATA_SOURCES
         .try_with(Arc::clone)
-        .map_err(|_| Error::NotSetDataSourcesError)
+        .map_err(|_| Error::NotSetDataSources)
 }