+++ src/lib.rs
@@ -1,12 +1,12 @@
 #![doc = include_str!("../README.md")]
 
-pub(crate) const DEFAULT_DATA_SOURCE_NAME: &str = "default";
+pub const DEFAULT_DATA_SOURCE: &str = "default";
 
 tokio::task_local! {
-    pub(crate) static DATA_SOURCES: std::sync::Arc<DataSources>;
+    pub static DATA_SOURCES: std::sync::Arc<DataSources>;
 }
 
-mod connection;
+mod data_source;
 mod data_sources;
 mod error;
 mod function;
@@ -14,10 +14,5 @@ mod middleware;
 mod transaction;
 
 pub use self::{
-    connection::Connection,
-    data_sources::DataSources,
-    error::Error,
-    function::{commit, current_txn, data_sources, default_txn, new_txn, rollback},
-    middleware::{SeaOrmEndpoint, SeaOrmMiddleware},
-    transaction::Transaction,
+    data_source::*, data_sources::*, error::*, function::*, middleware::*, transaction::*,
 };