+++ src/middleware.rs
@@ -1,16 +1,25 @@
-use std::sync::Arc;
+use std::{collections::HashMap, sync::Arc};
 
 use poem::{Endpoint, IntoResponse, Middleware, Request, Response, Result};
+use sea_orm::DatabaseConnection;
 
-use crate::{DataSources, DATA_SOURCES};
+use crate::{DataSources, DATA_SOURCES, DEFAULT_DATA_SOURCE};
 
 pub struct SeaOrmMiddleware {
-    data_sources: DataSources,
+    data_sources: HashMap<Arc<str>, DatabaseConnection>,
 }
 
 impl SeaOrmMiddleware {
-    pub fn new(data_sources: DataSources) -> Self {
+    pub fn with_default(conn: DatabaseConnection) -> Self {
+        let mut data_sources = HashMap::new();
+
+        data_sources.insert(Arc::<str>::from(DEFAULT_DATA_SOURCE), conn);
+
         Self { data_sources }
+    }
+
+    pub fn new(map: HashMap<Arc<str>, DatabaseConnection>) -> Self {
+        Self { data_sources: map }
     }
 }
 
@@ -26,7 +35,7 @@ impl<E: Endpoint> Middleware<E> for SeaOrmMiddleware {
 }
 
 pub struct SeaOrmEndpoint<E> {
-    data_sources: DataSources,
+    data_sources: HashMap<Arc<str>, DatabaseConnection>,
     inner: E,
 }
 
@@ -34,7 +43,7 @@ impl<E: Endpoint> Endpoint for SeaOrmEndpoint<E> {
     type Output = Response;
 
     async fn call(&self, req: Request) -> Result<Self::Output> {
-        let data_sources = Arc::new(self.data_sources.clone());
+        let data_sources = Arc::new(DataSources::new(&self.data_sources));
 
         let result = DATA_SOURCES
             .scope(data_sources.clone(), async { self.inner.call(req).await })